feat(ISV-6605): Fix static-tests error message on incorrect replaces version
