+++ operatorcert/static_tests/community/bundle.py
@@ -321,60 +321,6 @@ def check_api_version_constraints(bundle: Bundle) -> Iterator[CheckResult]:
         )
 
 
-@skip_fbc
-def check_replaces_availability(bundle: Bundle) -> Iterator[CheckResult]:
-    """
-    Check if the current bundle and the replaced bundle support the same OCP versions
-
-    Args:
-        bundle (Bundle): Operator bundle
-
-    Yields:
-        Iterator[CheckResult]: Failure if the version of the replaced bundle
-        does not match with the current bundle
-    """
-
-    replaces = bundle.csv.get("spec", {}).get("replaces")
-    if not replaces:
-        return
-    delimiter = ".v" if ".v" in replaces else "."
-    replaces_version = replaces.split(delimiter, 1)[1]
-
-    ver_to_dir = {
-        x.csv_operator_version: x.operator_version
-        for x in bundle.operator.all_bundles()
-    }
-    replaces_bundle = bundle.operator.bundle(ver_to_dir[replaces_version])
-
-    ocp_versions_str = bundle.annotations.get("com.redhat.openshift.versions")
-    replaces_ocp_version_str = replaces_bundle.annotations.get(
-        "com.redhat.openshift.versions"
-    )
-    if ocp_versions_str == replaces_ocp_version_str:
-        # The annotations match, no need to check further
-        return
-    organization = bundle.operator.repo.config.get("organization")
-
-    indexes = set(utils.get_ocp_supported_versions(organization, ocp_versions_str))
-    replaces_indexes = set(
-        utils.get_ocp_supported_versions(organization, replaces_ocp_version_str)
-    )
-
-    if indexes - replaces_indexes == set():
-        # The replaces bundle supports all the same versions as the current bundle
-        return
-    yield Fail(
-        f"Replaces bundle {replaces_bundle} {sorted(replaces_indexes)} does not support "
-        f"the same OCP versions as bundle {bundle} {sorted(indexes)}. In order to fix this issue, "
-        "align the OCP version range to match the range of the replaced bundle. "
-        "This can be done by setting the `com.redhat.openshift.versions` annotation in the "
-        "`metadata/annotations.yaml` file.\n"
-        f"`{bundle}` - `{ocp_versions_str}`\n"
-        f"`{replaces_bundle}` - `{replaces_ocp_version_str}`"
-    )
-    yield from []
-
-
 NON_FBC_SUGGESTION = (
     "[File Based Catalog (FBC)]"
     "(https://github.com/redhat-openshift-ecosystem/community-operators-prod/"
